+++ src/content.js
@@ -1,33 +1,5 @@
-// import { DiscountServices } from './common.js'
 console.log('Content: script is running!')
 let ignoreDomain
-
-const DiscountServices = { // eslint-disable-line no-unused-vars
-  forbrugsforeningen: {
-    name: 'Forbrugsforeningen',
-    arrayName: 'array_forbrugsforeningen',
-    databaseURL: 'https://cdn.jsdelivr.net/gh/zinen/rabatten-scraper@latest/dist/forbrugsforeningen.json',
-    homepage: 'http://forbrugsforeningen.dk'
-  },
-  logbuy: {
-    name: 'LogBuy',
-    arrayName: 'array_logbuy',
-    databaseURL: 'https://cdn.jsdelivr.net/gh/zinen/rabatten-scraper@latest/dist/logbuy.json',
-    homepage: 'http://logbuy.com'
-  },
-  coop: {
-    name: 'Coop partnerfordele',
-    arrayName: 'array_coop',
-    databaseURL: 'https://cdn.jsdelivr.net/gh/zinen/rabatten-scraper@latest/dist/coop.json',
-    homepage: 'https://partnerfordele.coop.dk/'
-  },
-  aeld: {
-    name: 'Ældre Sagen',
-    arrayName: 'array_aeld',
-    databaseURL: 'https://cdn.jsdelivr.net/gh/zinen/rabatten-scraper@latest/dist/aeld.json',
-    homepage: 'https://www.aeldresagen.dk/'
-  }
-}
 
 let tabDomainName = document.domain.split('.')
 if (tabDomainName[tabDomainName.length - 1] === 'uk') {
